+++ database/migrations/2019_12_10_143412_create_expenses_table.php
@@ -18,7 +18,7 @@ class CreateExpensesTable extends Migration
             $table->unsignedBigInteger('type_id');
             $table->double('amount');
             $table->Date('date_of_expenditure');
-            $table->integer('recorded_by')->unsigned();
+            $table->unsignedBigInteger('recorded_by')->unsigned();
             $table->timestamps();
 
             $table->foreign('type_id')->references('id')->on('expense_types')->onDelete('cascade');
