+++ database/migrations/2023_01_03_223756_create_guests_table.php
@@ -27,7 +27,7 @@ class CreateGuestsTable extends Migration
             $table->string('passport_number')->nullable();
             $table->string('nin')->nullable();
             $table->text('other_details')->nullable();
-            $table->integer('created_by')->unsigned();
+             $table->unsignedBigInteger('created_by')->unsigned();
             $table->timestamps();
             $table->foreign('created_by')->references('id')->on('staff');
         });
