+++ database/migrations/2023_01_09_151604_create_orders_table.php
@@ -24,7 +24,7 @@ class CreateOrdersTable extends Migration
             $table->unsignedBigInteger('guest_id')->nullable();
             $table->enum('status', Helper::getKitchenOrderStatuses());
             $table->timestamp('order_date')->default(DB::raw('CURRENT_TIMESTAMP'));
-            $table->integer('created_by')->unsigned();
+             $table->unsignedBigInteger('created_by')->unsigned();
             $table->timestamps();
 
             $table->foreign('room_id')->references('id')->on('rooms')->onDelete('set null');
