+++ database/migrations/2023_01_16_213322_create_staff_permissions_table.php
@@ -17,7 +17,7 @@ class CreateStaffPermissionsTable extends Migration
 
             $table->id();
             $table->unsignedBigInteger('permission_id');
-            $table->integer('staff_id')->unsigned();
+            $table->unsignedBigInteger('staff_id')->unsigned();
             $table->boolean('active')->default(0);
             $table->timestamps();
